pyproject.toml
[project]
name = "AVTrace"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Viktor Freiman", email = "freiman.viktor@gmail.com"},
]
dependencies = ["PyYAML>=6.0.2"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
